Improve documentation of Day 4 question: Number Complement

+++ May-LeetCoding-Challenge/04-Number-Complement/Number-Complement.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int findComplement(int num) {
-        long long i = 1;
+        long i = 1;
         while (i <= num) i <<= 1;
         return (i-1) ^ num;
     }
